Quiesce lint with some type handling.  Does not change anything.

+++ mdoc_html.c
@@ -830,6 +830,7 @@ mdoc_bx_pre(MDOC_ARGS)
 	return(0);
 }
 
+/* ARGSUSED */
 static int
 mdoc_it_pre(MDOC_ARGS)
 {
@@ -944,7 +945,7 @@ mdoc_it_pre(MDOC_ARGS)
 static int
 mdoc_bl_pre(MDOC_ARGS)
 {
-	size_t		 i;
+	int		 i;
 	struct htmlpair	 tag[3];
 	struct roffsu	 su;
 	char		 buf[BUFSIZ];
@@ -966,10 +967,10 @@ mdoc_bl_pre(MDOC_ARGS)
 		 * screen and we want to preserve that behaviour.
 		 */
 
-		for (i = 0; i < n->norm->Bl.ncols; i++) {
+		for (i = 0; i < (int)n->norm->Bl.ncols; i++) {
 			a2width(n->norm->Bl.cols[i], &su);
 			bufinit(h);
-			if (i < n->norm->Bl.ncols - 1)
+			if (i < (int)n->norm->Bl.ncols - 1)
 				bufcat_su(h, "width", &su);
 			else
 				bufcat_su(h, "min-width", &su);
